pyproject.toml
[project]
name = "django-pgconstraints"
version = "0.1.0a1"
description = "Declarative PostgreSQL constraint triggers for Django"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
keywords = ["django", "postgresql", "constraints", "triggers"]
authors = [{ name = "Oliver Haas", email = "ohaas@e1plus.de" }]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Framework :: Django",
  "Framework :: Django :: 5.2",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Typing :: Typed",
]
dependencies = ["Django>=5.2,<7", "django-pgtrigger>=4.17.0"]

[project.urls]
Homepage = "https://github.com/oliverhaas/django-pgconstraints"
Documentation = "https://oliverhaas.github.io/django-pgconstraints/"
Repository = "https://github.com/oliverhaas/django-pgconstraints.git"
Changelog = "https://oliverhaas.github.io/django-pgconstraints/reference/changelog/"

[dependency-groups]
dev = [
  "mypy==1.20.0",
  "django-stubs==6.0.2",
  "pre-commit==4.5.1",
  "pytest==9.0.2",
  "pytest-cov==7.1.0",
  "pytest-django==4.12.0",
  "pytest-flakefinder==1.1.0",
  "pytest-xdist==3.8.0",
  "ruff==0.15.9",
  "psycopg[c]==3.2.9",
  "testcontainers[postgres]==4.14.1",
]
docs = ["mkdocs==1.6.1", "mkdocs-material==9.7.6", "mike==2.1.4"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["django_pgconstraints"]

[tool.hatch.build.targets.sdist]
include = ["django_pgconstraints", "LICENSE", "README.md"]

[tool.ruff]
target-version = "py313"
line-length = 120
fix = true
extend-exclude = ["examples"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812", "D", "E501", "EM", "S608", "TRY003"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  "ANN",
  "ARG",
  "DJ008",
  "F841",
  "FBT",
  "PLR2004",
  "PT006",
  "PT011",
  "PT013",
  "PT018",
  "RUF012",
  "RUF059",
  "S101",
  "S105",
  "S108",
  "SLF001",
]

[tool.mypy]
python_version = "3.12"
mypy_path = "tests"
plugins = ["mypy_django_plugin.main"]
pretty = true
show_error_codes = true
show_error_context = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.django-stubs]
django_settings_module = "tests.settings.base"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings.base"
pythonpath = ["tests"]
addopts = "--cov=django_pgconstraints --cov-report=term-missing --cov-report=xml --no-cov-on-fail"
testpaths = ["tests"]
xfail_strict = true

[tool.coverage.run]
omit = ["tests/*"]

[tool.coverage.report]
precision = 2
skip_covered = true
